requirements.txt
transformers
torch
fastapi
uvicorn
pydantic
Pillow
starlette
lru_cache
py-hot-reload
optimum